requirements.txt
paho-mqtt~=1.6.1
influxdb_client~=1.39.0